# starter-kits/track-1-api-blueprint/openapi.yaml
openapi: 3.1.0
info:
  title: Backend Forge API - API Blueprint Track
  version: 1.0.0
  description: REST API with OpenAPI spec, Zod validation, and proper error handling
servers:
  - url: http://localhost:3000
paths:
  /api/v1/products:
    get:
      summary: List products with pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '201':
          description: Product created
        '400':
          description: Validation error
  /api/v1/products/{id}:
    get:
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product found
        '404':
          description: Product not found
    put:
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProduct'
      responses:
        '200':
          description: Product updated
        '404':
          description: Product not found
    delete:
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product deleted
        '404':
          description: Product not found
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateProduct:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        price:
          type: number
          exclusiveMinimum: 0
        description:
          type: string
          maxLength: 1000
    UpdateProduct:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        price:
          type: number
          exclusiveMinimum: 0
        description:
          type: string
          maxLength: 1000
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer